requirements.txt
boto3
requests==2.32.2
aiohttp==3.9.5
aiofiles==23.2.1
loguru==0.7.2
pydantic==2.18.2
python-dotenv==1.0.1
uvicorn==0.30.0
fastapi==0.111.0
fastapi-cli==0.0.4
